refactor: Update import paths and remove unused code in storage configuration
