--- src/ai/backend/storage/config/__init__.py
@@ -1,8 +1,0 @@
-"""
-Storage proxy configuration package.
-"""
-
-from .loaders import load_local_config, load_shared_config
-from .unified import StorageProxyUnifiedConfig
-
-__all__ = ["load_local_config", "load_shared_config", "StorageProxyUnifiedConfig"]
