+++ src/ai/backend/storage/config/loaders.py
@@ -5,12 +5,10 @@ import sys
 from pathlib import Path
 from pprint import pformat
 
-from ai.backend.common.config import (
-    override_key,
-    override_with_env,
-    read_from_file,
-)
+from ai.backend.common.config import ConfigurationError as BaseConfigError
+from ai.backend.common.config import override_key, override_with_env, read_from_file
 from ai.backend.common.etcd import AsyncEtcd, ConfigScopes
+from ai.backend.common.types import HostPortPair as CommonHostPortPair
 
 from .unified import StorageProxyUnifiedConfig
 
@@ -39,12 +37,10 @@ def load_local_config(config_path: Path | None, debug: bool = False) -> StorageP
             file=sys.stderr,
         )
         print(pformat(raw_cfg), file=sys.stderr)
-        from ai.backend.common.config import ConfigurationError as BaseConfigError
-
         raise BaseConfigError(raw_cfg) from e
 
 
-def load_shared_config(local_config: StorageProxyUnifiedConfig) -> AsyncEtcd:
+def make_etcd(local_config: StorageProxyUnifiedConfig) -> AsyncEtcd:
     """Load shared configuration from etcd."""
     etcd_credentials = None
     if local_config.etcd.user and local_config.etcd.password:
@@ -57,8 +53,6 @@ def load_shared_config(local_config: StorageProxyUnifiedConfig) -> AsyncEtcd:
         ConfigScopes.GLOBAL: "",
         ConfigScopes.NODE: f"nodes/storage/{local_config.storage_proxy.node_id}",
     }
-
-    from ai.backend.common.types import HostPortPair as CommonHostPortPair
 
     # Convert to common HostPortPair
     addr = local_config.etcd.addr
